Move gear calls generation functions to separate modules
